Restructure and cleanup
